add styling to codebuild output
